fixed that battered tree back to 4d11 see object 20: a small tree weight 40 4d8 -4 hitroll Kind of a special weapon, because maybe 1 or 2 race/classes in the mud can use it because of its weight. Battered tree: weight 60 4d11 -4 hitroll, +3 dam
It's nearly identical in stats to the foraged small tree.  really just there to make the mobs wielding it hit harder.  Not really considered a standard 'weapon' nor is small tree. damage range of small tree: 4-32, damage range of battered tree 4-44.  If you insist that this item is too strong, then would opt to make it !rent and transient and keep it at this strong for its intended function of keeping mob strong.