Allow _listening_ to gcc, acc, and petition in lockers/silent.
You can't talk to your magical thingy, but receiving should be ok.

It serves no gameplay purpose as you can already nchat or beep.  On the
other hand, it's a major annoyance as people randomly lose parts of a
conversation, without a notice to them or those speaking.  And, even if
we wanted to plug the nchat/beep holes, there's Discord...